interchange: refactor Avro Encoder
Refactors the Avro Encoder used for Kafka sinks to more easily add
additional Encoder types in the future.
