+++ src/interchange/src/avro/encode.rs
@@ -20,6 +20,8 @@ use repr::adt::numeric::{self, NUMERIC_AGG_MAX_PRECISION, NUMERIC_DATUM_MAX_PREC
 use repr::{ColumnName, ColumnType, Datum, RelationDesc, Row, ScalarType};
 use serde_json::json;
 
+use crate::encode::Encode;
+use crate::json::build_row_schema_json;
 use mz_avro::types::{DecimalValue, Value};
 use mz_avro::Schema;
 
@@ -120,22 +122,14 @@ fn encode_message_unchecked(
     buf
 }
 
-/// Manages encoding of Avro-encoded bytes.
-pub struct Encoder {
+/// Generates key and value Avro schemas
+pub struct AvroSchemaGenerator {
     value_columns: Vec<(ColumnName, ColumnType)>,
     key_info: Option<KeyInfo>,
     writer_schema: Schema,
 }
 
-impl fmt::Debug for Encoder {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_struct("Encoder")
-            .field("writer_schema", &self.writer_schema)
-            .finish()
-    }
-}
-
-impl Encoder {
+impl AvroSchemaGenerator {
     pub fn new(
         key_desc: Option<RelationDesc>,
         value_desc: RelationDesc,
@@ -175,7 +169,7 @@ impl Encoder {
                 columns,
             }
         });
-        Encoder {
+        AvroSchemaGenerator {
             value_columns,
             key_info,
             writer_schema,
@@ -199,17 +193,68 @@ impl Encoder {
             .as_ref()
             .map(|KeyInfo { columns, .. }| columns.as_slice())
     }
+}
+
+impl fmt::Debug for AvroSchemaGenerator {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        f.debug_struct("SchemaGenerator")
+            .field("writer_schema", &self.writer_schema)
+            .finish()
+    }
+}
+
+/// Manages encoding of Avro-encoded bytes.
+pub struct AvroEncoder {
+    schema_generator: AvroSchemaGenerator,
+    key_schema_id: Option<i32>,
+    value_schema_id: i32,
+}
+
+impl fmt::Debug for AvroEncoder {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        f.debug_struct("AvroEncoder")
+            .field("writer_schema", &self.schema_generator.writer_schema)
+            .finish()
+    }
+}
+
+impl AvroEncoder {
+    pub fn new(
+        schema_generator: AvroSchemaGenerator,
+        key_schema_id: Option<i32>,
+        value_schema_id: i32,
+    ) -> Self {
+        AvroEncoder {
+            schema_generator,
+            key_schema_id,
+            value_schema_id,
+        }
+    }
 
     pub fn encode_key_unchecked(&self, schema_id: i32, row: Row) -> Vec<u8> {
-        let schema = self.key_writer_schema().unwrap();
-        let columns = self.key_columns().unwrap();
+        let schema = self.schema_generator.key_writer_schema().unwrap();
+        let columns = self.schema_generator.key_columns().unwrap();
         encode_message_unchecked(schema_id, row, schema, columns)
     }
 
     pub fn encode_value_unchecked(&self, schema_id: i32, row: Row) -> Vec<u8> {
-        let schema = self.value_writer_schema();
-        let columns = self.value_columns();
+        let schema = self.schema_generator.value_writer_schema();
+        let columns = self.schema_generator.value_columns();
         encode_message_unchecked(schema_id, row, schema, columns)
+    }
+}
+
+impl Encode for AvroEncoder {
+    fn get_format_name(&self) -> &str {
+        "avro"
+    }
+
+    fn encode_key_unchecked(&self, row: Row) -> Vec<u8> {
+        self.encode_key_unchecked(self.key_schema_id.unwrap(), row)
+    }
+
+    fn encode_value_unchecked(&self, row: Row) -> Vec<u8> {
+        self.encode_value_unchecked(self.value_schema_id, row)
     }
 }
 
@@ -474,144 +519,4 @@ pub fn encode_debezium_transaction_unchecked(
     debug_assert!(avro.validate(DEBEZIUM_TRANSACTION_SCHEMA.top_node()));
     mz_avro::encode_unchecked(&avro, &DEBEZIUM_TRANSACTION_SCHEMA, &mut buf);
     buf
-}
-
-fn build_row_schema_field<F: FnMut() -> String>(
-    namer: &mut F,
-    names_seen: &mut HashSet<String>,
-    typ: &ColumnType,
-) -> serde_json::value::Value {
-    let mut field_type = match &typ.scalar_type {
-        ScalarType::Bool => json!("boolean"),
-        ScalarType::Int16 | ScalarType::Int32 | ScalarType::Oid => json!("int"),
-        ScalarType::Int64 => json!("long"),
-        ScalarType::Float32 => json!("float"),
-        ScalarType::Float64 => json!("double"),
-        ScalarType::Date => json!({
-            "type": "int",
-            "logicalType": "date",
-        }),
-        ScalarType::Time => json!({
-            "type": "long",
-            "logicalType": "time-micros",
-        }),
-        ScalarType::Timestamp | ScalarType::TimestampTz => json!({
-            "type": "long",
-            "logicalType": "timestamp-micros"
-        }),
-        ScalarType::Interval => json!({
-            "type": "fixed",
-            "size": 12,
-            "logicalType": "duration"
-        }),
-        ScalarType::Bytes => json!("bytes"),
-        ScalarType::String => json!("string"),
-        ScalarType::Jsonb => json!({
-            "type": "string",
-            "connect.name": "io.debezium.data.Json",
-        }),
-        ScalarType::Uuid => json!({
-            "type": "string",
-            "logicalType": "uuid",
-        }),
-        ScalarType::Array(element_type) | ScalarType::List { element_type, .. } => {
-            let inner = build_row_schema_field(
-                namer,
-                names_seen,
-                &ColumnType {
-                    nullable: true,
-                    scalar_type: (**element_type).clone(),
-                },
-            );
-            json!({
-                "type": "array",
-                "items": inner
-            })
-        }
-        ScalarType::Map { value_type, .. } => {
-            let inner = build_row_schema_field(
-                namer,
-                names_seen,
-                &ColumnType {
-                    nullable: true,
-                    scalar_type: (**value_type).clone(),
-                },
-            );
-            json!({
-                "type": "map",
-                "values": inner
-            })
-        }
-        ScalarType::Record {
-            fields,
-            custom_name,
-            ..
-        } => {
-            let (name, name_seen) = match custom_name {
-                Some(name) => (name.clone(), !names_seen.insert(name.clone())),
-                None => (namer(), false),
-            };
-            if name_seen {
-                json!(name)
-            } else {
-                let fields = fields.to_vec();
-                let json_fields = build_row_schema_fields(&fields, names_seen, namer);
-                json!({
-                    "type": "record",
-                    "name": name,
-                    "fields": json_fields
-                })
-            }
-        }
-        ScalarType::Numeric { scale } => {
-            let (p, s) = match scale {
-                Some(scale) => (NUMERIC_DATUM_MAX_PRECISION, usize::from(*scale)),
-                None => (NUMERIC_AGG_MAX_PRECISION, NUMERIC_DATUM_MAX_PRECISION),
-            };
-            json!({
-                "type": "bytes",
-                "logicalType": "decimal",
-                "precision": p,
-                "scale": s,
-            })
-        }
-    };
-    if typ.nullable {
-        field_type = json!(["null", field_type]);
-    }
-    field_type
-}
-
-pub(super) fn build_row_schema_fields<F: FnMut() -> String>(
-    columns: &[(ColumnName, ColumnType)],
-    names_seen: &mut HashSet<String>,
-    namer: &mut F,
-) -> Vec<serde_json::value::Value> {
-    let mut fields = Vec::new();
-    for (name, typ) in columns.iter() {
-        let field_type = build_row_schema_field(namer, names_seen, typ);
-        fields.push(json!({
-            "name": name,
-            "type": field_type,
-        }));
-    }
-    fields
-}
-
-/// Builds the JSON for the row schema, which can be independently useful.
-pub(super) fn build_row_schema_json(
-    columns: &[(ColumnName, ColumnType)],
-    name: &str,
-) -> serde_json::value::Value {
-    let mut name_idx = 0;
-    let fields = build_row_schema_fields(columns, &mut Default::default(), &mut move || {
-        let ret = format!("com.materialize.sink.record{}", name_idx);
-        name_idx += 1;
-        ret
-    });
-    json!({
-        "type": "record",
-        "fields": fields,
-        "name": name
-    })
 }
